+++ website/forms/__init__.py
@@ -1,0 +1,3 @@
+from website.forms.payment_type_form import PaymentTypeForm
+from website.forms.product_form import ProductForm
+from website.forms.user_form import UserForm
