+++ website/forms/product_form.py
@@ -1,0 +1,8 @@
+from django import forms
+from website.models.product_model import Product
+
+class ProductForm(forms.ModelForm):
+
+    class Meta:
+        model = Product
+        fields = ('title', 'description', 'price', 'quantity',)
