+++ website/forms/user_form.py
@@ -8,9 +8,3 @@ class UserForm(forms.ModelForm):
     class Meta:
         model = User
         fields = ('username', 'email', 'password', 'first_name', 'last_name',)
-
-class ProductForm(forms.ModelForm):
-
-    class Meta:
-        model = Product
-        fields = ('title', 'description', 'price', 'quantity',)
