+++ website/models/__init__.py
@@ -1,0 +1,5 @@
+from website.models.product_order_model import ProductOrder
+from website.models.product_category_model import Category
+from website.models.product_model import Product
+from website.models.order_model import Order
+from website.models.payment_type_model import PaymentType
