+++ website/views/__init__.py
@@ -1,0 +1,6 @@
+from website.views.index_view import index
+from website.views.list_products_view import list_products
+from website.views.login_user_view import login_user
+from website.views.register_view import register
+from website.views.sell_product_view import sell_product
+from website.views.user_logout_view import user_logout
