+++ website/views/index_view.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+def index(request):
+    template_name = 'index.html'
+    return render(request, template_name, {})
