+++ website/views/login_user_view.py
@@ -1,0 +1,35 @@
+from django.template import RequestContext
+from django.contrib.auth import login, authenticate
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.shortcuts import render
+
+def login_user(request):
+    '''Handles the creation of a new user for authentication
+
+    Method arguments:
+      request -- The full HTTP request object
+    '''
+
+    # Obtain the context for the user's request.
+    context = RequestContext(request)
+
+    # If the request is a HTTP POST, try to pull out the relevant information.
+    if request.method == 'POST':
+
+        # Use the built-in authenticate method to verify
+        username=request.POST['username']
+        password=request.POST['password']
+        authenticated_user = authenticate(username=username, password=password)
+
+        # If authentication was successful, log the user in
+        if authenticated_user is not None:
+            login(request=request, user=authenticated_user)
+            return HttpResponseRedirect('/')
+
+        else:
+            # Bad login details were provided. So we can't log the user in.
+            print("Invalid login details: {}, {}".format(username, password))
+            return HttpResponse("Invalid login details supplied.")
+
+
+    return render(request, 'login.html', {}, context)
